Move `flatten_buffer` to `_utils`
This ran into cyclic import issues in `pickle_compat`. So move
`flatten_buffer` to its own module free of these issues.
