+++ pandas/tests/io/test_pickle.py
@@ -39,7 +39,7 @@ from pandas.compat import (
     is_platform_little_endian,
 )
 from pandas.compat._optional import import_optional_dependency
-from pandas.compat.pickle import flatten_buffer
+from pandas.compat._utils import flatten_buffer
 import pandas.util._test_decorators as td
 
 import pandas as pd
